requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==43.0.1
Django==5.1.1
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
Markdown==3.7
mysqlclient==2.2.4
packaging==24.1
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
